Changed ingredient model back to old one with server data filling in

+++ iOS/Food-Doods-Prototype/Food-Doods-Prototype/Controllers/Item.swift
@@ -13,11 +13,11 @@ struct Item {
     var name: String
     var image: UIImage?
     var location: FoodLocation
-    var amount: Int
+    var amount: Double
     var expiresIn: Int //In days
     var shelfLife: Int //In days
     
-    init(name: String, image: UIImage?, location: FoodLocation, amount: Int, expires: Int, shelfLife: Int) {
+    init(name: String, image: UIImage?, location: FoodLocation, amount: Double, expires: Int, shelfLife: Int) {
         self.name = name
         self.image = image
         self.location = location
